Update README.md
Cbind function may be used only to demonstrate the Matrix since the numbers are displayed as characters in ("") which are not processed in R for statistical analysis. I tried to do mean function for the cbind function results and got this error: In mean.default(Results[, 2]) : argument is not numeric or logical: returning NA.
I also received the same warning when tried to do the mean function using this code: mean(results.df[,2:3]). I tried mean(results.df[,2]) and it calculated the mean for the second column, and mean(results.df[,3])also for the third column, however, the combination didn't work. So I tried other function in order to find the means from selected columns in a data frame, so I found colMeans function which enables us to put desired columns in function.